[refactor] Drop leading ``_`` from ``compare_text.py`` and ``highlight.py``
These two modules are imported by ``raises.py`` (outside the
``_pytest.assertion`` package), so the leading underscore did not
match the existing convention in the package (``rewrite.py``,
``truncate.py``, ``util.py``). The genuinely-package-private modules
(``_compare_set``, ``_compare_mapping``, ``_compare_sequence``,
``_compare_any``, ``_guards``, ``_typing``) keep the underscore.
