+++ src/_pytest/assertion/highlight.py
